sign-out handle missing cart

+++ lib/actions/user.actions.ts
@@ -45,7 +45,12 @@ export async function signInWithCredentials(
 export async function signOutUser() {
   // get current users cart and delete it so it does not persist to next user
   const currentCart = await getMyCart();
-  await prisma.cart.delete({ where: { id: currentCart?.id } });
+
+  if (currentCart?.id) {
+    await prisma.cart.delete({ where: { id: currentCart?.id } });
+  } else {
+    console.warn('No Cart found for deletion');
+  }
   await signOut();
 }
